Start work on course gallery
