+++ package.json
@@ -25,6 +25,7 @@
     "node-fetch": "^3.2.0",
     "prism-react-renderer": "^1.1.1",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.4.0",
     "react-dom": "^17.0.2",
     "react-keyed-flatten-children": "^1.3.0",
     "rehype-katex": "^6.0.2",
