+++ src/components/Header.module.scss
@@ -4,3 +4,17 @@
 .header {
   background-color: $blue-800;
 }
+
+.nav-link {
+  color: rgba(255, 255, 255, 1);
+  padding: 0.5rem 1rem;
+  text-decoration: none;
+}
+
+.nav-link:hover {
+  color: $yellow;
+}
+
+.nav-link.active {
+  color: $yellow;
+}
